feat : implement login and register functionality

+++ app/Http/Controllers/NoteController.php
@@ -14,7 +14,8 @@ class NoteController extends Controller
     {
 
         // fetching 
-        $notes = Note::query()->orderBy("created_at", "desc")->paginate(15);
+        // where to filter notes based on user
+        $notes = Note::query()->where('user_id', request()->user()->id)->orderBy("created_at", "desc")->paginate(15);
 
         // Dumb and Die : Dumbs the data and immediately ends script execution.
         // dd($notes);
@@ -42,30 +43,39 @@ class NoteController extends Controller
         ]);
 
         // hard coding user_id
-        $data['user_id'] = 1;
-
+        $data['user_id'] = $request->user()->id;
         // returns the note data saved in the DB.
         $note = Note::create($data);
+        
 
         // redirecting to show page with $note data to display the newly created note.
         // message : session message -> can be used in the layout file
-        return to_route('note.show', $note)->with('notification', 'Not was created.');
+        return to_route('note.show', $note)->with('notification', 'Note was created.');
     }
 
     /**
      * Display the specified resource.
      */
     // $note will be accessed from the url.
-    public function show(Note $note)
-    {
+    public function show($id) {
+
+        $note = Note::find($id);
+        if($note->user_id !== request()->user()->id){
+            // for bidden error
+            abort(403);
+        }
         return view('note.show', ['note' => $note]);
     }
 
     /**
      * Show the form for editing the specified resource.
      */
-    public function edit(Note $note)
-    {
+    public function edit($id) {
+        $note = Note::find($id);
+        if($note->user_id !== request()->user()->id){
+            // for bidden error
+            abort(403);
+        }
         return view('note.edit', ['note' => $note]);
     }
 
@@ -75,10 +85,16 @@ class NoteController extends Controller
     // we will get the $note object passed from UI 
     public function update(Request $request, Note $note)
     {
+
+        if($note->user_id !== request()->user()->id){
+            // for bidden error
+            abort(403);
+        }
          // server side validation -> body contains the note data
          $data = $request->validate([
             'note' => ['required', 'string']
         ]);
+
 
 
         // invoking the update method to update the
@@ -87,7 +103,7 @@ class NoteController extends Controller
 
         // redirecting to show page with $note data to display the newly created note.
         // message : session message -> can be used in the layout file
-        return to_route('note.show', $note)->with('notification', 'Not was updated.');
+        return to_route('note.show', $note)->with('notification', 'Note was updated.');
     }
 
     /**
@@ -95,8 +111,12 @@ class NoteController extends Controller
      */
     public function destroy(Note $note)
     {
+        if($note->user_id !== request()->user()->id){
+            // for bidden error
+            abort(403);
+        }
         $note->delete();
 
-        return to_route('note.index')->with("notification", "Not was deleted");
+        return to_route('note.index')->with("notification", "Note was deleted");
     }
 }
